Enhance stage select UI/navigation; fix score cast
Add background panels and chapter navigation for stage select: introduce three back Texture2D objects (m_mode_in_{L,M,R}_back), toggle their alpha on selection, and implement ChangeStageSelectScene(houkou) to handle chapter changes and update display textures. Adjust selection input logic to call ChangeStageSelectScene at chapter boundaries and update select index accordingly. Fix UI scaling by changing horizontal scale multiplier from 2.0 to 1.7778 for L/M/R icons to preserve aspect ratio. Minor gameplay fixes: in Stage10, cast score accumulation to int (Score += static_cast<int>(baseScore * ScoreMultiplier)); in Stage0, change scene transition target from STAGE_SELECT to MODE_SELECT. Add new asset assets/texture/stageselect/back.png and update StageSelectScene.h with new members and the ChangeStageSelectScene declaration.

+++ DX11_hew2/Stage10.cpp
@@ -353,7 +353,7 @@ void Stage10::GameUpdate()
 					enemy->StartMayuing(centroid);
 					++eliminatedCount;
 					ScoreMultiplier += 0.1f;	//1体目は1倍、以降0.1倍ずつ倍率アップ
-					Score += baseScore * ScoreMultiplier;	//100*スコア倍率
+					Score += static_cast<int>(baseScore * ScoreMultiplier);	//100*スコア倍率
 					if (Score >= 99999)
 					{
 						Score = 99999; //スコアカンスト
